Moved helper scripts in their folder + updated import path
