+++ helpers/trendings_helper.py
